requirements.txt
python-dotenv
ipykernel
ipywidgets
requests
numpy
pandas
scipy
scikit-learn
matplotlib
torch
transformers
tqdm
openai
groq
gradio
langchain
langchain-core
langchain-text-splitters
langchain-openai
langchain-chroma
langchain-community
datasets==3.6.0
matplotlib
google-generativeai
anthropic
chromadb
plotly
jupyter-dash
beautifulsoup4
pydub
modal
ollama
psutil
setuptools
speedtest-cli
sentence_transformers
feedparser
protobuf==3.20.2
wandb
